+++ .github/workflows/build.yaml
@@ -30,6 +30,12 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
+      - name: Configure AWS Creds
+        uses: aws-actions/configure-aws-credentials@v2
+        with:
+          aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY }}
+          aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+          aws-region: us-west-2
       - name: Check isort, black, and flake8
         run: |
           pip install black flake8 isort
@@ -40,9 +46,17 @@ jobs:
       - name: Install project dependencies
         run: poetry install --no-interaction
       - name: Run tests with coverage
+        env:
+          API_USER: ${{ secrets.API_USER }}
+          API_PASSWORD: ${{ secrets.API_PASSWORD }}
+          BUCKET_NAME: ${{ secrets.BUCKET_NAME }}
+          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY }}
+          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
         run: |
+          export TEST_WORKDIR=`pwd`/`date +%s`
+          mkdir $TEST_WORKDIR
           # Run the unit tests
-          poetry run python3 -m coverage run -a -m unittest discover -v src
+          poetry run python3 -m coverage run -a -m unittest discover -v src/krkn_lib_kubernetes/tests/
           # Generate the coverage HTML report
           poetry run python3 -m coverage html
       - name: Publish coverage report to job summary
