+++ pyproject.toml
@@ -19,6 +19,7 @@ kubeconfig = "^1.1.1"
 
 [tool.poetry.group.test.dependencies]
 jinja2 = "^3.1.2"
+boto3 = "^1.28.12"
 
 [build-system]
 requires = ["poetry-core>=1.1.0"]
